refactor: precedingService 불필요함으로 삭제
