+++ surbear-api/src/main/java/com/surbear/example/service/ExampleService.java
@@ -1,7 +1,10 @@
 package com.surbear.example.service;
 
+import com.surbear.mock.example.entity.example.ExampleEntity;
+import com.surbear.mock.example.mapper.ExampleMapper;
 import com.surbear.mock.example.model.Example;
-import com.surbear.mock.example.service.ExamplePrecedingService;
+import com.surbear.mock.example.repository.ExampleRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,10 +13,24 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 @Service
 public class ExampleService {
-    private final ExamplePrecedingService service;
+    private final ExampleRepository repository;
+    private final ExampleMapper mapper;
 
     @Transactional
-    public Long saveExample(Example example){
-        return service.create(example);
+    public Long saveExample(Example example) {
+        return create(example);
+    }
+
+
+    private Long create(Example example) {
+        ExampleEntity exampleEntity = mapper.toEntity(example);
+        return repository.save(exampleEntity).getId();
+    }
+
+    private Example get(Long id) {
+        var entity = repository.findById(id)
+                .orElseThrow(EntityNotFoundException::new);
+
+        return mapper.toModel(entity);
     }
 }
